requirements.txt
appdirs==1.4.4
argcomplete==3.0.8
asttokens==2.2.1
backcall==0.2.0
decorator==5.1.1
escpos==1.9
executing==1.2.0
future==0.18.3
ipython==8.13.2
jedi==0.18.2
matplotlib-inline==0.1.6
mysql-connector-python==8.0.33
numpy==1.24.3
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
protobuf==3.20.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pypng==0.20220715.0
pyserial==3.5
python-barcode==0.14.0
python-dateutil==2.8.2
pytz==2023.3
pyusb==1.2.1
PyYAML==6.0
qrcode==7.4.2
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.6.0
tzdata==2023.3
viivakoodi==0.8.0
wcwidth==0.2.6
